chore: convert to typescript
