+++ block.json
@@ -1,5 +1,4 @@
 {
   "version": "1.0",
-  "frontendEntry": "./frontend/index.js",
-  "frontendTestingEntry": "./test/index.js"
+  "frontendEntry": "./frontend/index.tsx"
 }
