+++ frontend/app.tsx
@@ -8,16 +8,24 @@ import {
   colors,
 } from '@airtable/blocks/ui'
 import React from 'react'
-import GenerateRecordForm from './form'
+import GenerateRecordForm from './components/form'
 
-const RandomRecordGeneratorApp = () => {
+/**
+ * The main app component. Fetches the currently selected table using
+ * the cursor, checks access to create records in the table, then
+ * renders the generate record form.
+ * @returns
+ */
+const RandomRecordGeneratorApp: React.FC = () => {
   const base = useBase()
   const cursor = useCursor()
   const activeTable = cursor.activeTableId
   const table = activeTable && base.getTableByIdIfExists(activeTable)
+
+  // There is no active table.
   if (!table) {
     return (
-      <Box padding="1rem">
+      <Box padding={2}>
         <Heading size="xlarge">Random Record Generator</Heading>
         <Text textColor={colorUtils.getHexForColor(colors.GRAY_BRIGHT)}>
           Select a table to get started.
@@ -25,21 +33,28 @@ const RandomRecordGeneratorApp = () => {
       </Box>
     )
   }
+
   const checkTablePermission = table.checkPermissionsForCreateRecord()
 
   return (
-    <Box width="90%" margin="1.5rem auto">
+    <Box width="90%" marginY={2} marginX="auto">
       {table && (
         <>
-          <Text>
+          <Text variant="paragraph">
             Generate random records for <strong>{table.name}</strong>.{' '}
           </Text>
 
           {checkTablePermission.hasPermission ? (
             <GenerateRecordForm table={table} />
           ) : (
-            <Text textColor={colorUtils.getHexForColor(colors.RED_DARK_1)}>
-              {checkTablePermission.reasonDisplayString}
+            <Text
+              variant="paragraph"
+              textColor={colorUtils.getHexForColor(colors.RED_DARK_1)}
+            >
+              {
+                // @ts-ignore
+                checkTablePermission.reasonDisplayString
+              }
             </Text>
           )}
         </>
