+++ frontend/components/results.tsx
@@ -1,10 +1,26 @@
 import { Box, Heading, Text, ProgressBar, Button } from '@airtable/blocks/ui'
 import React from 'react'
 
-const GenerateResults = ({ generated, numberOfRecords, onDone }) => (
+interface Props {
+  generated: number
+  numberOfRecords: number
+  onDone: () => void
+}
+
+/**
+ * The results component. Displays the number of records generated and
+ * a progress bar. Calls onDone when the user clicks a Start Over button
+ * once record generation is complete.
+ * @returns
+ */
+const GenerateResults: React.FC<Props> = ({
+  generated,
+  numberOfRecords,
+  onDone,
+}) => (
   <Box marginTop="1rem">
     <Heading>Generating records</Heading>
-    <Text>
+    <Text variant="paragraph">
       Generated <strong>{generated}</strong> of{' '}
       <strong>{numberOfRecords}</strong> records.
     </Text>
