+++ frontend/index.tsx
@@ -1,6 +1,6 @@
 import { initializeBlock } from '@airtable/blocks/ui'
 import React from 'react'
 
-import RandomRecordGeneratorApp from './random-name-generator'
+import RandomRecordGeneratorApp from './app'
 
 initializeBlock(() => <RandomRecordGeneratorApp />)
