+++ frontend/lib/generators/generators/generator.d.ts
@@ -1,0 +1,10 @@
+import { FieldType, Field } from '@airtable/blocks/models'
+
+interface Generator {
+  id: string
+  name: string
+  types: FieldType[]
+  generate: (preview: boolean, field: any) => any
+}
+
+export default Generator
