+++ package.json
@@ -3,7 +3,7 @@
   "description": "Randomly generates Airtable records",
   "license": "Apache-2.0",
   "scripts": {
-    "lint": "eslint frontend test",
+    "lint": "eslint frontend",
     "test:jsdom": "jest --testMatch '**/*.jsdom.test.js'"
   },
   "dependencies": {
@@ -24,6 +24,8 @@
     "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "@testing-library/user-event": "^12.5.0",
+    "@typescript-eslint/eslint-plugin": "^5.41.0",
+    "@typescript-eslint/parser": "^5.41.0",
     "core-js": "^3.8.1",
     "cypress": "^6.2.0",
     "eslint": "^6.3.0",
@@ -32,6 +34,7 @@
     "geckodriver": "^2.0.2",
     "jest": "^24.9.0",
     "selenium-webdriver": "^4.0.0-alpha.8",
-    "start-server-and-test": "^1.11.6"
+    "start-server-and-test": "^1.11.6",
+    "typescript": "^4.8.4"
   }
 }
